Separate tagvalue controller from tagkey controller
